fixed cart functionality

+++ database/migrations/2024_01_12_134729_create_carts_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->integer('price');
             $table->text('details');
             $table->string('category');
-            $table->tinyInteger('quantity')->default('1')->nullable();
+            $table->integer('quantity')->default('1')->nullable();
             $table->bigInteger('user_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users');
             $table->timestamps();
